+++ CapstonProjectBE/Controllers/PostController.cs
@@ -1,8 +1,11 @@
 ﻿using Application.IService;
 using Application.ViewModels.PostDTO;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Authorization.Infrastructure;
 using Microsoft.AspNetCore.Cors.Infrastructure;
 using Microsoft.AspNetCore.Mvc;
+using System.Drawing.Printing;
 
 namespace CapstonProjectBE.Controllers
 {
@@ -12,12 +15,14 @@ namespace CapstonProjectBE.Controllers
     public class PostController : Controller
     {
         private readonly IPostService _postService;
-        public PostController(IPostService postService)
+        private readonly IAuthenService _authenService;
+        public PostController(IPostService postService, IAuthenService authenService)
         {
             _postService = postService;
+            _authenService = authenService;
         }
 
-        [Authorize]
+        [Authorize(Roles = "Customer")]
         [HttpPost]
         public async Task<IActionResult> CreatePost(CreatePostDTO createPostDTO)
         {
@@ -31,23 +36,51 @@ namespace CapstonProjectBE.Controllers
         }
 
         [Authorize]
-        [HttpGet]
+        [HttpGet("user")]
         public async Task<IActionResult> GetPostsByUserId(int userId)
         {
-            var result = await _postService.GetPostsByUserId(userId);
+            var user = await _authenService.GetUserByTokenAsync(HttpContext.User);
+            var result = await _postService.GetPostsByProjectId(userId, user.UserId);
             if (!result.Success)
             {
                 return BadRequest(result);
             }
-
             return Ok(result);
         }
 
-        [Authorize]
-        [HttpGet("pagination")]
+        [HttpGet("pagination/user")]
+        [AllowAnonymous]
         public async Task<IActionResult> GetPaginatedPostsByUserId(int userId, int page = 1, int pageSize = 20)
         {
-            var result = await _postService.GetPaginatedPostsByUserId(userId, page, pageSize);
+            var user = await _authenService.GetUserByTokenAsync(HttpContext.User);
+            var result = await _postService.GetPaginatedPostsByProjectId(userId, page, pageSize, user.UserId);
+            if (!result.Success)
+            {
+                return BadRequest(result);
+            }
+
+            return Ok(result);
+        }
+
+        [HttpGet("project")]
+        [AllowAnonymous]
+        public async Task<IActionResult> GetPostsByProjectId(int projectId)
+        {
+            var user = await _authenService.GetUserByTokenAsync(HttpContext.User);
+            var result = user == null ? await _postService.GetPostsByProjectId(projectId, null) : await _postService.GetPostsByProjectId(projectId, user.UserId);
+            if (!result.Success)
+            {
+                return BadRequest(result);
+            }
+            return Ok(result);
+        }
+
+        [HttpGet("pagination/project")]
+        [AllowAnonymous]
+        public async Task<IActionResult> GetPaginatedPostsByProjectId(int projectId, int page = 1, int pageSize = 20)
+        {
+            var user = await _authenService.GetUserByTokenAsync(HttpContext.User);
+            var result = user == null ? await _postService.GetPaginatedPostsByProjectId(projectId, page, pageSize, null) : await _postService.GetPaginatedPostsByProjectId(projectId, page, pageSize, user.UserId);
             if (!result.Success)
             {
                 return BadRequest(result);
@@ -57,10 +90,19 @@ namespace CapstonProjectBE.Controllers
         }
 
 
-        [Authorize]
+        [Authorize(Roles = "Customer")]
         [HttpPut]
         public async Task<IActionResult> UpdatePost(int postId, CreatePostDTO createPostDTO)
         {
+            var user = await _authenService.GetUserByTokenAsync(HttpContext.User);
+            if (user == null)
+            {
+                return Unauthorized();
+            }
+            if (user.Role == "Customer" && !(await _postService.CheckIfPostHasUserId(postId, user.UserId)))
+            {
+                return Forbid();
+            }
             var result = await _postService.UpdatePost(postId, createPostDTO);
             if (!result.Success)
             {
@@ -70,10 +112,19 @@ namespace CapstonProjectBE.Controllers
             return Ok(result);
         }
 
-        [Authorize]
+        [Authorize(Roles = "Customer, Staff")]
         [HttpDelete]
         public async Task<IActionResult> RemovePost(int postId)
         {
+            var user = await _authenService.GetUserByTokenAsync(HttpContext.User);
+            if (user == null)
+            {
+                return Unauthorized();
+            }
+            if (user.Role == "Customer" && !(await _postService.CheckIfPostHasUserId(postId, user.UserId)))
+            {
+                return Forbid();
+            }
             var result = await _postService.RemovePost(postId);
             if (!result.Success)
             {
